refactor: improve type safety and API ergonomics

+++ src/models/candle.rs
@@ -1,9 +1,10 @@
 //! OHLC candlestick channel models.
 
+use rust_decimal::Decimal;
 use serde::Deserialize;
 
 /// An update message from the `ohlc` (candles) channel.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct CandleUpdateResponse {
     pub channel: String,
     #[serde(rename = "type")]
@@ -13,17 +14,17 @@ pub struct CandleUpdateResponse {
 }
 
 /// A single OHLC candlestick bar.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct CandleData {
     pub symbol: String,
-    pub open: f64,
-    pub high: f64,
-    pub low: f64,
-    pub close: f64,
+    pub open: Decimal,
+    pub high: Decimal,
+    pub low: Decimal,
+    pub close: Decimal,
     /// Volume-weighted average price for this candle.
-    pub vwap: f64,
+    pub vwap: Decimal,
     pub trades: u64,
-    pub volume: f64,
+    pub volume: Decimal,
     /// Start timestamp of this candle's time window.
     pub interval_begin: String,
     /// Candle duration in minutes.
