+++ src/models/orders.rs
@@ -1,9 +1,10 @@
 //! Level-3 (individual orders) channel models.
 
+use rust_decimal::Decimal;
 use serde::Deserialize;
 
 /// An update message from the `level3` (orders) channel.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct OrdersUpdateResponse {
     pub channel: String,
     #[serde(rename = "type")]
@@ -12,7 +13,7 @@ pub struct OrdersUpdateResponse {
 }
 
 /// Level-3 order book data for a single trading pair.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct OrdersData {
     pub symbol: String,
     pub bids: Vec<OrderEntry>,
@@ -22,13 +23,13 @@ pub struct OrdersData {
 }
 
 /// A single order in the level-3 book.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct OrderEntry {
     /// Lifecycle event for this order: `"add"`, `"modify"`, or `"delete"`.
     /// `None` on the initial snapshot.
     pub event: Option<String>,
     pub order_id: String,
-    pub limit_price: f64,
-    pub order_qty: f64,
+    pub limit_price: Decimal,
+    pub order_qty: Decimal,
     pub timestamp: String,
 }
